generate_video_with_bone/requirements.txt
# Function dependencies, for example:
# package>=version
protobuf>=3.20.3
google-cloud-firestore>=2.11.0
google-cloud-storage>=2.9.0
opencv-python==4.7.0.72
functions-framework>=3.4.0
pytz>=2023.3
google-events>=0.7.0
cloudevents>=1.9.0